check token before exicute function

+++ src/Controller/CartController.php
@@ -7,6 +7,8 @@ use App\Model\Entity\Cart;
 use App\Model\Entity\CartProduct;
 use App\Model\Table\CartProductsTable;
 use App\Model\Table\CartTable;
+use App\Model\Entity\User;
+use Cake\I18n\Time;
 
 /**
  * Cart Controller
@@ -249,33 +251,40 @@ class CartController extends AppController {
 			// $data=$this->request->data();//cart_id,product_id,qty,type[default-1]
 			$product_id = $this->request->data ( 'product_id' );
 			$product_qty = $this->request->data ( 'qty' );
-			if ($product_id != null && $product_qty != null) {
-				$cart_id = $this->__getCartId ();
-				if (! $this->__isInCart ( $cart_id, $product_id, 1 )) {
-					$data = [ 
-							'cart_id' => $cart_id,
-							'product_id' => $product_id,
-							'qty' => $product_qty,
-							'type' => 1 
-					];
-					
-					$cart_product_model = $this->loadModel ( 'CartProducts' );
-					$product_entity = $cart_product_model->newEntity ( $data );
-					$saving = $cart_product_model->save ( $product_entity );
-					if ($saving) {
-						$return ['status'] = 0;
-						$return ['message'] = 'Pruduct is added to catr';
+			$token = $this->request->data ( 'token' );
+			$chck = $this->__checkToken ( $token );
+			if ($chck ['boolean']) {
+				if ($product_id != null && $product_qty != null) {
+					$cart_id = $this->__getCartId ();
+					if (! $this->__isInCart ( $cart_id, $product_id, 1 )) {
+						$data = [ 
+								'cart_id' => $cart_id,
+								'product_id' => $product_id,
+								'qty' => $product_qty,
+								'type' => 1 
+						];
+						
+						$cart_product_model = $this->loadModel ( 'CartProducts' );
+						$product_entity = $cart_product_model->newEntity ( $data );
+						$saving = $cart_product_model->save ( $product_entity );
+						if ($saving) {
+							$return ['status'] = 0;
+							$return ['message'] = 'Pruduct is added to catr';
+						} else {
+							$return ['status'] = 500;
+							$return ['message'] = 'Pruduct is not added to catr';
+						}
 					} else {
-						$return ['status'] = 500;
-						$return ['message'] = 'Pruduct is not added to catr';
+						$return ['status'] = 0;
+						$return ['message'] = 'The pruduct already in your catr';
 					}
 				} else {
-					$return ['status'] = 0;
-					$return ['message'] = 'The pruduct already in your catr';
+					$return ['status'] = 401;
+					$return ['message'] = "please select product to add cart";
 				}
 			} else {
-				$return ['status'] = 401;
-				$return ['message'] = "please select product to add cart";
+				$return ['status'] = 500;
+				$return ['message'] = $chck ['message'];
 			}
 		} else {
 			$return ['status'] = 500;
@@ -293,41 +302,48 @@ class CartController extends AppController {
 		header ( 'Content-type: application/json' );
 		if ($this->request->is ( 'post' )) {
 			$product_id = $this->request->data ( 'product_id' );
-			if ($product_id != null) {
-				$cart_id = $this->__getCurrentCartId ();
-				if ($cart_id) {
-					$cart_product_model = $this->loadModel ( 'CartProducts' );
-					
-					$product = $cart_product_model->find ( 'all', [ 
-							'fields' => [ 
-									'id' 
-							],
-							'conditions' => [ 
-									'cart_id' => $cart_id,
-									'product_id' => $product_id,
-									'type' => 1 
-							] 
-					] )->toArray ();
-					if (sizeof ( $product ) > 0) {
-						if ($cart_product_model->delete ( $cart_product_model->get ( $product [sizeof ( $product ) - 1]->id ) )) {
-							$return ['status'] = 0;
-							$return ['message'] = 'Pruduct deleted successfully';
-							$return ['result'] = $this->__getcartIn();
+			$token = $this->request->data ( 'token' );
+			$chck = $this->__checkToken ( $token );
+			if ($chck ['boolean']) {
+				if ($product_id != null) {
+					$cart_id = $this->__getCurrentCartId ();
+					if ($cart_id) {
+						$cart_product_model = $this->loadModel ( 'CartProducts' );
+						
+						$product = $cart_product_model->find ( 'all', [ 
+								'fields' => [ 
+										'id' 
+								],
+								'conditions' => [ 
+										'cart_id' => $cart_id,
+										'product_id' => $product_id,
+										'type' => 1 
+								] 
+						] )->toArray ();
+						if (sizeof ( $product ) > 0) {
+							if ($cart_product_model->delete ( $cart_product_model->get ( $product [sizeof ( $product ) - 1]->id ) )) {
+								$return ['status'] = 0;
+								$return ['message'] = 'Pruduct deleted successfully';
+								$return ['result'] = $this->__getcartIn ();
+							} else {
+								$return ['status'] = 500;
+								$return ['message'] = 'Culd not delete the product';
+							}
 						} else {
 							$return ['status'] = 500;
-							$return ['message'] = 'Culd not delete the product';
+							$return ['message'] = 'The product not found in the cart';
 						}
 					} else {
 						$return ['status'] = 500;
-						$return ['message'] = 'The product not found in the cart';
+						$return ['message'] = 'you havent create a cart';
 					}
 				} else {
 					$return ['status'] = 500;
-					$return ['message'] = 'you havent create a cart';
+					$return ['message'] = 'please select product id';
 				}
 			} else {
 				$return ['status'] = 500;
-				$return ['message'] = 'please select product id';
+				$return ['message'] = $chck ['message'];
 			}
 		} else {
 			$return ['status'] = 500;
@@ -341,23 +357,32 @@ class CartController extends AppController {
 				'post' 
 		] );
 		header ( 'Content-type: application/json' );
-		$cart_id = $this->__getCurrentCartId ();
-		if ($cart_id) {
-			$cart_product_model = $this->loadModel ( 'CartProducts' );
-			if ($cart_product_model->deleteAll ( [ 
-					'cart_id' => $cart_id,
-					'type' => 1 
-			] )) {
-				$return ['status'] = 0;
-				$return ['message'] = 'cart clear success';
+		$token = $this->request->data ( 'token' );
+		$chck = $this->__checkToken ( $token );
+		if ($chck ['boolean']) {
+			
+			$cart_id = $this->__getCurrentCartId ();
+			if ($cart_id) {
+				$cart_product_model = $this->loadModel ( 'CartProducts' );
+				if ($cart_product_model->deleteAll ( [ 
+						'cart_id' => $cart_id,
+						'type' => 1 
+				] )) {
+					$return ['status'] = 0;
+					$return ['message'] = 'cart clear success';
+				} else {
+					$return ['status'] = 500;
+					$return ['message'] = 'cartclear not clear or car is empty';
+				}
 			} else {
 				$return ['status'] = 500;
-				$return ['message'] = 'cartclear not clear or car is empty';
+				$return ['message'] = 'you havent create a cart';
 			}
 		} else {
 			$return ['status'] = 500;
-			$return ['message'] = 'you havent create a cart';
+			$return ['message'] = $chck ['message'];
 		}
+		
 		echo json_encode ( $return );
 		die ();
 	}
@@ -386,7 +411,7 @@ class CartController extends AppController {
 						if ($cart_product_model->save ( $product )) {
 							$return ['status'] = 0;
 							$return ['message'] = 'Pruduct qty updated successfully';
-							$return ['result'] = $this->__getcartIn();
+							$return ['result'] = $this->__getcartIn ();
 						} else {
 							$return ['status'] = 500;
 							$return ['message'] = 'Culd not update the qty';
@@ -412,52 +437,61 @@ class CartController extends AppController {
 	}
 	public function getcart() {
 		$this->request->allowMethod ( [ 
-				'post'
+				'post' 
+		] );
+		header ( 'Content-type: application/json' );
+		
+		$token = $this->request->data ( 'token' );
+		$chck = $this->__checkToken ( $token );
+		if ($chck ['boolean']) {
+			
+			$cart_id = $this->__getCurrentCartId ();
+			
+			if ($cart_id) {
+				
+				$total = $this->__getTotal ( $cart_id );
+				$cart_products = CartProductsTable::getCart ( $cart_id, 1 );
+				
+				if (sizeof ( $cart_products ) > 0) {
+					$return ['status'] = 0;
+					$return ['message'] = 'success';
+					$return ['result'] ['product_list'] = $cart_products;
+					$return ['result'] ['total'] = $total;
+				} else {
+					$return ['status'] = 0;
+					$return ['message'] = 'your cart is empty';
+					$return ['result'] = [ ];
+				}
+			} else {
+				$return ['status'] = 500;
+				$return ['message'] = "you haven't create a cart";
+			}
+		} else {
+			$return ['status'] = 500;
+			$return ['message'] = $chck ['message'];
+		}
+		
+		echo json_encode ( $return );
+		die ();
+	}
+	public function __getcartIn() {
+		$this->request->allowMethod ( [ 
+				'post',
+				'get' 
 		] );
 		header ( 'Content-type: application/json' );
 		$cart_id = $this->__getCurrentCartId ();
 		
 		if ($cart_id) {
 			
-			$total=$this->__getTotal($cart_id);
-			$cart_products = CartProductsTable::getCart($cart_id,1);
+			$total = $this->__getTotal ( $cart_id );
+			$cart_products = CartProductsTable::getCart ( $cart_id, 1 );
 			
 			if (sizeof ( $cart_products ) > 0) {
 				$return ['status'] = 0;
 				$return ['message'] = 'success';
-				$return ['result']['product_list'] = $cart_products;
-				$return ['result']['total'] = $total;
-			} else {
-				$return ['status'] = 0;
-				$return ['message'] = 'your cart is empty';
-				$return ['result'] = [ ];
-			} 
-		} else {
-			$return ['status'] = 500;
-			$return ['message'] = "you haven't create a cart";
-		}
-		echo json_encode ( $return );
-		die ();
-	}
-	
-	public function __getcartIn() {
-		$this->request->allowMethod ( [
-				'post',
-				'get'
-		] );
-		header ( 'Content-type: application/json' );
-		$cart_id = $this->__getCurrentCartId ();
-	
-		if ($cart_id) {
-				
-			$total=$this->__getTotal($cart_id);
-			$cart_products = CartProductsTable::getCart($cart_id,1);
-				
-			if (sizeof ( $cart_products ) > 0) {
-				$return ['status'] = 0;
-				$return ['message'] = 'success';
-				$return ['result']['product_list'] = $cart_products;
-				$return ['result']['total'] = $total;
+				$return ['result'] ['product_list'] = $cart_products;
+				$return ['result'] ['total'] = $total;
 			} else {
 				$return ['status'] = 0;
 				$return ['message'] = 'your cart is empty';
@@ -469,23 +503,68 @@ class CartController extends AppController {
 		}
 		return $return;
 		die ();
-	}	
-	
+	}
 	public function __getTotal($cart_id) {
-		$tax_p=0;//tax persontage 10
-		$discount_p=0;//discount persentage 5
-		$counpon_value=0;//call to a function to find coupon values
-		$sub_total=CartTable::getTotal($cart_id, 1);
+		$tax_p = 0; // tax persontage 10
+		$discount_p = 0; // discount persentage 5
+		$counpon_value = 0; // call to a function to find coupon values
+		$sub_total = CartTable::getTotal ( $cart_id, 1 );
 		
-	
-		$tax=$sub_total*$tax_p/100;
-		$discount=$sub_total*$discount_p/100;
-		$grand_total=$sub_total+$tax-$discount-$counpon_value;
-	
+		$tax = $sub_total * $tax_p / 100;
+		$discount = $sub_total * $discount_p / 100;
+		$grand_total = $sub_total + $tax - $discount - $counpon_value;
+		
 		$total ['sub_total'] = $sub_total;
 		$total ['tax'] = $tax;
 		$total ['discount'] = $discount;
 		$total ['grand_total'] = $grand_total;
 		return $total;
+	}
+	private function __updateMobTokenTime($token) {
+		$user_model = $this->loadModel ( 'users' );
+		$user = $user_model->find ( 'all', [ 
+				'conditions' => [ 
+						'mobtoken' => $token 
+				] 
+		] );
+		$user->mobtoken_created_at = date ( 'Y-m-d H:i:s' );
+		return $user_model->save ( $user );
+	}
+	/**
+	 *
+	 * @param unknown $token        	
+	 * @return multitype:boolean string
+	 */
+	function __checkToken($token) {
+		$user_model = $this->loadModel ( 'users' );
+		$user = $user_model->find ( 'all', [ 
+				'conditions' => [ 
+						'mobtoken' => $token 
+				] 
+		] )->first ();
+		if (sizeof ( $user ) <= 0) {
+			return [ 
+					'boolean' => false,
+					'message' => 'token not found' 
+			];
+		} else {
+			$mobtoken_created_at = $user->mobtoken_created_at;
+			$mobtoken_created_at = new Time ( $mobtoken_created_at );
+			if ($mobtoken_created_at->wasWithinLast ( '720mins' )) {
+				$user->mobtoken_created_at = date ( 'Y-m-d H:i:s' );
+				$user_model->save ( $user );
+				
+				return [ 
+						'boolean' => true,
+						'message' => 'token matched',
+						'user_id' =>$user->id 
+				];
+			} else {
+				return [ 
+						'boolean' => false,
+						'message' => 'token expired' 
+				];
+			}
+		}
 	}
 }
